pyproject.toml
[project]
name = "multi-agentic-rag"
version = "1.0.0"
description = "Production multi-agent RAG system with streaming"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",

    # LLM providers
    "openai>=1.14.0",
    "anthropic>=0.21.0",

    # Vector DB
    "qdrant-client>=1.8.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.28",
    "asyncpg>=0.29.0",

    # Document parsing
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.2",
    "tiktoken>=0.6.0",

    # Gmail API
    "google-api-python-client>=2.120.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",

    # HTTP / async
    "aiohttp>=3.9.3",
    "httpx>=0.27.0",

    # Config / utils
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]
